Azure_Functions/requirements.txt
azure-functions
requests==2.24.0